Allow arrays of points to be passed through directly to rasterize_to_tri_strip interface.
Using the PathBuilder interface can be slow when we need to also access the canonical
representation of the path for use-cases such as hashing outside wpf-gpu-raster. Also,
it allows for faster construction of the point arrays when we can reuse or have specific
knowledge of the path.

This does still keep the PathBuilder interface around as a simpler abstraction over the
more fundamental rasterize_to_tri_strip function for use-cases where we just need to
build the representation once and be done with it.

This also compresses the edge type arrays to only have 1 byte per group of associated
points, rather than 1 type per every point, so that we now behave more like Skia where
a type corresponds to a "verb" of sorts.