Allow PathBuilder to be reset whilst reusing Vec storages
Adds a PathBuilder.reset() method along with a C binding that resets a
PathBuilder to its default state whilst persisting the storage of the
points and types Vecs. The consumer can then reuse a PathBuilder
multiple times whilst reducing expensive reallocations.

A consequence of this is that get_path() now copies the contents of
the Vecs in to a boxed slice as opposed to consuming the Vecs. In
practice, Vec::into_box_slice() usually results in a realloc anyway
due to shrinking the Vec's storage to the used size.